requirements.txt
# Core RAG dependencies
pypdf>=3.0.0
langdetect>=1.0.9
requests>=2.28.0

# ColBERT retrieval
colbert-ai>=0.2.19
faiss-cpu>=1.7.0

# MaxSimE for explainability
maxsime @ git+https://github.com/ebritoc/MaxSimE.git

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Web UI
gradio>=4.0.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-timeout>=2.1.0

# Optional (progress bars, nicer output)
tqdm>=4.65.0
